requirements.txt

beautifulsoup4==4.10.0
Bottleneck==1.3.2
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2021.10.8
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
firebase-admin==5.3.0
Flask==2.2.2
fonttools==4.25.0
gensim==4.1.2
gensim-sum-ext==0.1.2
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
itsdangerous==2.1.2
joblib==1.1.0


Keras-Preprocessing==1.1.2

lxml==4.8.0
MarkupSafe==2.1.1

msgpack==1.0.4
munkres==1.1.4

nltk==3.3.0
numpy==1.21.4

pandas==1.3.5

Pillow==8.4.0

proto-plus==1.22.1
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycorenlp==0.3.0

pyspellchecker==0.6.3
python-crfsuite==0.9.7
pytz==2021.3
regex==2022.1.18

rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0

sklearn-crfsuite==0.3.6

soupsieve==2.3.1

spacy-wordnet==0.0.5
tabulate==0.8.9

threadpoolctl==3.1.0

uritemplate==4.1.1

webencodings==0.5.1
Werkzeug==2.2.2
wikipedia==1.4.0
